+++ packages/searchkit-apollo-resolvers/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-canary.20](https://github.com/searchkit/searchkit/compare/v3.0.0-canary.19...v3.0.0-canary.20) (2020-12-13)
+
+**Note:** Version bump only for package @searchkit/apollo-resolvers
+
+
+
+
+
 # [3.0.0-canary.19](https://github.com/searchkit/searchkit/compare/v3.0.0-canary.12...v3.0.0-canary.19) (2020-12-07)