+++ packages/searchkit-client/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-canary.20](https://github.com/searchkit/searchkit/compare/v3.0.0-canary.19...v3.0.0-canary.20) (2020-12-13)
+
+
+### Features
+
+* allow query to be adjusted ([407384f](https://github.com/searchkit/searchkit/commit/407384fea64c6a7c2e7651934d7ae080798734ce)), closes [#806](https://github.com/searchkit/searchkit/issues/806)
+
+
+
+
+
 # [3.0.0-canary.19](https://github.com/searchkit/searchkit/compare/v3.0.0-canary.12...v3.0.0-canary.19) (2020-12-07)